bstr: impl From<T> instead of AsRef
Currently there is no way of creating a bstr from a Vec<u8> without
doing a copy of the entire slice.

Implement all From<T>'s we are converting from instead of using
AsRef<[u8]>. This way is a bit more work but will increase performance
by not performing copies in some cases.