transaction/multipart: htp_header_t ptr to owned
Remove the name and value pointers in htp_header_t so that it owns its
parts.

Also change the htp_headers_t list type so that all the elements are
owned.

Also store this headers list in the transaction as owned instead of ptrs.

Also replaces the table functions. Adds convenience methods on
request_headers and response_headers instead of having to go through the
htp_table_t interface directly. The main issue is that we changed the
type of htp_headers_t and functions like htp_table_get_c() cast away the
type safety of the elements.