decompressors: refactor module
Replaces the htp_decompressors module with a new idiomatic decompressors
module.

Uses the std::io::Write trait to chain decompressors and then call a
callback. This tries to be as compatible as possible with the previous
version and the external API remains the same.

Uses the new streaming decompressor in lzma-rs to handle the lzma
Content-Encoding.

Replaces the C implementation with a simple wrapper around the lzma-rs
crate because it is still required by suricata to do SWF LZMA
decompression.

Moves all config options related to decompression into the decompressors
module. This will keep all things related to decompression in a single
place.

Adds the Options struct which implements Default to specify configuration
options. Also changes the inner decompressor to handle the lzma memlimit,
with a value of 0 disabling lzma altogether.

Adds test cases for configuration options (memlimit, bomb_limit,
time_limit).