connp: Replace in_tx and out_tx with Transactions
Transactions ensures that there is always a valid request and response
transaction, which eliminates much of the boilerplate sanity checking.
It also means that the transaction index no longer needs to be
manipulated directly in ConnectionParser, since Transactions just has
next() convenience functions to increment the current request or
response transactions.

Also add an API to flush incomplete transactions, meant to be used when a caller is dropping a connection and wants to get callbacks for everything that has not yet completed.