table: When doing a key lookup, search the 'trimmed' key name.
This lets the programmer just specify the header name like usual,
without having to worry about whether or not the header on the wire
may have had leading or trailing whitespace.

The returned value still returns the 'on the wire' representation,
so any leading or trailing whitespace is present in  the returned
value, but the lookup function ignores leading and trailing whitespace.