Change the request state to be in progress before callbacks.
If the receiver of the callback wants to know "how many
transactions are there" it may get the wrong answer if the
state of the tx is still not started.

Fixes one of the suricata tests which does just this.

In the longer term, using the number of transactions as a
proxy for the "active transaction number" is a hack, and we
should fix the users of this library to stop doing it because
the transaction already knows what number it is and the callers
should just use that directly.