response: Add test for clearing buffered data.
Follows C version commit fbfd1479d6e98141023621d1145d715542326360

Here we just add the test, because the rust version was already handling
the buffer correctly. But we were incorrectly calling chomp() on the
input line assuming it was a response line, but then treating it
as body data in this corner case. So preserve the input line throughout
and chomp() the response message when we have determined it is actually
a response message.