cbindgen: Add option for headers output directory.
When this library is built with cbindgen support as a dependency,
the output location for the generated headers is hard for the parent
program to predict and fetch. So provide an environment variable
the caller can set beforehand to just specify the header output
directory.

eg: CBINDGEN_HEADERS_DIR=/path/to/headers cargo build ...