headers: Fix no name parsing with mixed line endings.
When parsing headers, force the parser to try both the normal
and the no-name/colon case for headers by requiring the normal
case to be complete (turning Err::Incomplete into Err::Error).

For the no-name/colon parser, switch it to just use the Value
parser, which is more correct and handles the difference between
request and response parsing (where the response parser accepts
a single CR as a line ending).