headers: Make header evasion detection more strict
http evader 274 hides a header in what looks like wrapping:

X-Foo: bar
 Transfer-Encoding: chunked
Yet-another-header: foo

To some clients is looks like wrapping, but some clients will read the
Transfer-Encoding header and apply chunked encoding.

We check for this condition by detecting token chars and a ":" in the
wrapped value, and then treating it like a new header. This change
further restricts the token chars part to be not-empty, meaning input
like this:

X-Foo: bar
 : chunked
Yet-another-header: foo

Gets treated as wrapping instead of a new header with empty name.